fixed tests and reorganized tests folder
