--- tests/test_webserver/test_webserver_api_trade.py
@@ -1,58 +1,0 @@
-
-from fastapi import FastAPI
-from fastapi import status
-from fastapi.testclient import TestClient
-
-from limit_order_book_webserver.fastapi_webserver import app
-from lib_financial_exchange.financial_exchange_types import OrderId
-from lib_financial_exchange.exceptions import DuplicateOrderIdError
-
-from tests.test_webserver.helper import helper_generate_order_without_order_id
-from tests.test_webserver.helper import helper_generate_order_id
-
-client = TestClient(app)
-
-
-def test_order_trade():
-    json = helper_generate_order_without_order_id('PYTH', 'BUY', 1000, 10)
-    response = client.post('/api/send_order', json=json)
-    assert response.status_code == 200
-    assert response.json() == {
-        'status': 'success',
-        'message': None,
-        'order_id': 1,
-        'trades': []
-    }
-
-    json = helper_generate_order_without_order_id('PYTH', 'SELL', 995, 5)
-    response = client.post('/api/send_order', json=json)
-    assert response.status_code == 200
-    assert response.json() == {
-        'status': 'success',
-        'message': None,
-        'order_id': 2,
-        'trades': [
-            {
-                'ticker': 'PYTH',
-                'order_id_maker': 1,
-                'order_id_taker': 2,
-                'price': 995,
-                'volume': 5,
-            }
-        ]
-    }
-
-    json = helper_generate_order_id(1)
-    response = client.post('/api/cancel_order', json=json)
-    assert response.status_code == 200
-    assert response.json() == {
-        'status': 'success',
-        'message': 'order id 1 cancelled',
-        'order': {
-            'order_id': 1,
-            'ticker': 'PYTH',
-            'order_side': 'BUY',
-            'price': 1000,
-            'volume': 5,
-        }
-    }
